feat(): edit published ext
