+++ libs/design-system-components/src/components/ExtensionEditStep1Form/Header/index.tsx
@@ -20,9 +20,14 @@ import { getColorClasses } from '@akashaorg/design-system-core/lib/utils';
 import { useCloseActions } from '@akashaorg/design-system-core/lib/utils/useCloseActions';
 import { DeleteImageModal } from './DeleteImageModal';
 import { AkashaAppApplicationType } from '@akashaorg/typings/lib/sdk/graphql-types-new';
+import Pill from '@akashaorg/design-system-core/lib/components/Pill';
+import { capitalize } from 'lodash';
+import ExtensionIcon from '@akashaorg/design-system-core/lib/components/ExtensionIcon';
 
 export type HeaderProps = {
   extensionType?: AkashaAppApplicationType;
+  nsfw?: boolean;
+  showExtraInfo?: boolean;
   coverImage: Image;
   logoImage: Image;
   cancelLabel: string;
@@ -49,6 +54,8 @@ export type HeaderProps = {
 
 export const Header: React.FC<HeaderProps> = ({
   extensionType,
+  nsfw,
+  showExtraInfo,
   coverImage,
   logoImage,
   cancelLabel,
@@ -301,6 +308,23 @@ export const Header: React.FC<HeaderProps> = ({
           />
         </Stack>
       </Stack>
+      {showExtraInfo && (
+        <Stack direction="row" spacing="gap-2" justify="end" align="center">
+          <Pill
+            type="info"
+            label={capitalize(extensionType?.toLowerCase())}
+            icon={<ExtensionIcon type={extensionType} />}
+          />
+          {nsfw && (
+            <Pill
+              type="info"
+              label={'NSFW'}
+              background={{ light: 'errorFade', dark: 'errorDark' }}
+              color={{ light: 'errorDark', dark: 'white' }}
+            />
+          )}
+        </Stack>
+      )}
       <ImageModal
         show={showEditImage}
         title={appImageType === 'logo-image' ? imageTitle.logoImage : imageTitle.coverImage}
