+++ libs/design-system-core/src/components/Pill/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { ReactElement, useState } from 'react';
 import Button from '../Button';
 import Stack from '../Stack';
 import Text, { TextProps } from '../Text';
@@ -20,6 +20,7 @@ type ActionPillProps = Pick<
 
 export type InfoPillProps = Pick<TextProps, 'weight' | 'color' | 'customStyle'> & {
   label: string;
+  icon?: ReactElement;
   borderColor?: Color;
   background?: Color;
   customTextStyle?: string;
@@ -61,7 +62,8 @@ export type PillProps =
  **/
 const Pill: React.FC<PillProps> = props => {
   if (props.type === 'info') {
-    const { borderColor, background, color, label, weight, customStyle, customTextStyle } = props;
+    const { borderColor, background, color, icon, label, weight, customStyle, customTextStyle } =
+      props;
 
     const borderStyle = borderColor ? `border ${getColorClasses(borderColor, 'border')}` : '';
     return (
@@ -71,6 +73,7 @@ const Pill: React.FC<PillProps> = props => {
         background={background}
         customStyle={`min-h-min min-w-min rounded-full py-1 px-2 ${borderStyle} ${customStyle}`}
       >
+        {icon && icon}
         <Text
           variant="footnotes2"
           selectable={false}
