+++ libs/sdk/src/registry/registry-queries.graphql
@@ -6,6 +6,7 @@ fragment AkashaAppFragment on AkashaApp{
   name
   displayName
   keywords
+  nsfw
   createdAt
   releases(first: 5, sorting: {createdAt: DESC} ){
     edges {
